requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.8
django-appconf==1.0.4
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-googledrive-storage==1.5.2
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
Markdown==3.2.2
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
